move home illustration to a separate module
