+++ client/components/home/easy/configuration.tsx
@@ -1,7 +1,6 @@
-import clsx from "clsx";
 import { motion, useAnimation } from "framer-motion";
 import Image from "next/image";
-import { ComponentProps, FC, ReactNode, useEffect } from "react";
+import { FC, useEffect } from "react";
 import { useInView } from "react-intersection-observer";
 
 import useTheme from "~/client/hooks/theme";
@@ -12,52 +11,7 @@ import sectionLiveDark from "~/public/images/home/section-live-dark.svg";
 import sectionLiveLight from "~/public/images/home/section-live-light.svg";
 
 import ButtonLink from "../buttonLink";
-
-const Illustration: FC<{ parts: ReactNode[] } & ComponentProps<"div">> = ({
-  parts,
-  className,
-  ...rest
-}) => {
-  const animation = useAnimation();
-  const [ref, inView] = useInView({ threshold: 0.5 });
-  useEffect(() => {
-    if (inView) animation.start("visible");
-  }, [animation, inView]);
-  return (
-    <div ref={ref} className={clsx(className, "pointer-events-none relative pt-12")} {...rest}>
-      <div
-        className={clsx(
-          "md:hidden absolute w-px left-1/2 top-0 bottom-0 -translate-x-px bg-gradient-to-b",
-          "from-neutral-200 dark:from-neutral-800"
-        )}
-      />
-      <svg className="md:hidden absolute left-1/2 top-0 w-2 h-2 -translate-x-[4.5px] -translate-y-1 fill-neutral-200 dark:fill-neutral-800">
-        <circle cx="50%" cy="50%" r="50%" />
-      </svg>
-      {parts.map((part, index) => (
-        <motion.div
-          variants={{
-            visible: {
-              opacity: 1,
-              x: 0,
-              transition: { duration: 0.15, ease: "easeInOut", delay: index * 0.15 },
-            },
-            hidden: {
-              opacity: 0,
-              x: 50,
-              transition: { duration: 0.15, ease: "easeInOut", delay: index * 0.15 },
-            },
-          }}
-          initial="hidden"
-          animate={animation}
-          key={index}
-        >
-          {part}
-        </motion.div>
-      ))}
-    </div>
-  );
-};
+import Illustration from "../illustration";
 
 const EasyConfiguration: FC = () => {
   const theme = useTheme();
@@ -76,8 +30,8 @@ const EasyConfiguration: FC = () => {
           ref={ref}
           className="text-muted text-lg md:text-xl lg:text-2xl"
           variants={{
-            visible: { opacity: 1, transition: { duration: 0.15, ease: "easeInOut" } },
-            hidden: { opacity: 0, transition: { duration: 0.15, ease: "easeInOut" } },
+            visible: { opacity: 1, transition: { duration: 0.15 } },
+            hidden: { opacity: 0, transition: { duration: 0.15 } },
           }}
           initial="hidden"
           animate={animation}
